hash set doesn't have two same element.
so when to add a key to hash set, we need to check whether the key is
already in the set.
As with any hash, every item must have a unique key. Your application must enforce key uniqueness. Before you add an item to the hash table, you must first know (if in doubt, check!) that the key is not already in use. You can check whether a key already exists in the hash table using HASH_FIND.